- make EnsureMounted more generic and work for Filesystem
