+++ storage/Filesystems/FilesystemImpl.h
@@ -26,6 +26,7 @@
 
 
 #include "storage/Utils/Enum.h"
+#include "storage/Utils/FileUtils.h"
 #include "storage/Filesystems/Filesystem.h"
 #include "storage/Devices/DeviceImpl.h"
 
@@ -65,9 +66,24 @@ namespace storage
 
 	virtual void print(std::ostream& out) const override;
 
+	/**
+	 * A string useable for mounting the filesystem. E.g. for
+	 * BlkFilesystems the name of a device and for Nfs server plus
+	 * path. Used for messages and mount command.
+	 */
+	virtual string get_mount_string() const = 0;
+
+	/**
+	 * A string useable for mounting the filesystem. E.g. for
+	 * BlkFilesystems the name of a device and for Nfs server plus
+	 * path. Used for entries in /etc/fstab.
+	 */
+	virtual string get_mount_by_string() const = 0;
+
 	virtual Text do_mount_text(const string& mountpoint, Tense tense) const = 0;
 	virtual void do_mount(const Actiongraph::Impl& actiongraph, const string& mountpoint) const = 0;
 
+	const Filesystem* get_filesystem() const { return to_filesystem(get_device()); }
 	virtual Text do_umount_text(const string& mountpoint, Tense tense) const = 0;
 	virtual void do_umount(const Actiongraph::Impl& actiongraph, const string& mountpoint) const = 0;
 
@@ -163,6 +179,32 @@ namespace storage
 	};
 
     }
+
+
+    class EnsureMounted : boost::noncopyable
+    {
+
+    public:
+
+	/**
+	 * Ensures that the blk filesystem is mounted somewhere.
+	 *
+	 * The mode is not enforced.
+	 */
+	EnsureMounted(const Filesystem* filesystem);
+
+	/**
+	 * Returns any mountpoint of the filesystem.
+	 */
+	string get_any_mountpoint() const;
+
+    private:
+
+	const Filesystem* filesystem;
+
+	unique_ptr<TmpMount> tmp_mount;
+
+    };
 
 }
