+++ storage/Filesystems/VfatImpl.cc
@@ -61,7 +61,7 @@ namespace storage
     ContentInfo
     Vfat::Impl::detect_content_info_pure() const
     {
-	EnsureMounted ensure_mounted(this);
+	EnsureMounted ensure_mounted(get_filesystem());
 
 	ContentInfo content_info;
